save item

+++ src/AccountGoWeb/Models/Purchasing/Payment.cs
@@ -4,8 +4,8 @@
     {
         public int InvoiceId { get; set; }
         public int VendorId { get; set; }
-        public double InvoiceAmount { get; set; }
-        public double AmountToPay { get; set; }
+        public decimal InvoiceAmount { get; set; }
+        public decimal AmountToPay { get; set; }
         public int AccountId { get; set; }
         public System.DateTime Date { get; set; }
     }
